+++ internal/correlationid/generator.go
@@ -1,6 +1,9 @@
 package correlationid
 
-type IdGenerator interface {
+//go:generate mockery -name=Generator
+
+// Generator generates a correlation ID.
+type Generator interface {
 	// Generate generates a unique string which can be used as a correlation ID.
 	Generate() string
 }